Deal with NullStorage in AND.
