+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/NullStorage.java
@@ -64,8 +64,6 @@ public class NullStorage extends Storage<Void> {
     ops.add(new NullOp(Maps.DIV));
     ops.add(new NullOp(Maps.MOD));
     ops.add(new NullOp(Maps.POWER));
-
-    ops.add(new NullAndOp());
     ops.add(new NullOrOp());
 
     return ops;
@@ -188,21 +186,6 @@ public class NullStorage extends Storage<Void> {
         }
       }
       return builder.seal();
-    }
-  }
-
-  private static class NullAndOp extends BoolAndNullOp {
-    public NullAndOp() {
-      super(Maps.AND);
-    }
-
-    @Override
-    protected Boolean doBool(boolean a) {
-      if (a) {
-        return null;
-      } else {
-        return false;
-      }
     }
   }
